+++ domain/config.go
@@ -1,10 +1,21 @@
 package domain
 
 import (
+	"errors"
 	"fmt"
+	"reflect"
 
+	"github.com/mitchellh/mapstructure"
+	orderbookplugindomain "github.com/osmosis-labs/sqs/domain/orderbook/plugin"
 	passthroughdomain "github.com/osmosis-labs/sqs/domain/passthrough"
+	"github.com/spf13/viper"
 )
+
+// Plugin defines the interface for the plugins.
+type Plugin interface {
+	IsEnabled() bool
+	GetName() string
+}
 
 // Config defines the config for the sidecar query server.
 type Config struct {
@@ -48,6 +59,85 @@ type Config struct {
 	// SideCarQueryServer CORS configuration.
 	CORS *CORSConfig `mapstructure:"cors"`
 }
+
+// UnmarshalConfig handles the custom unmarshaling for the Config struct, particularly for Plugins.
+func UnmarshalConfig() (*Config, error) {
+	var config Config
+
+	// Use Viper's Unmarshal method to decode the configuration, except for the Plugins field
+	if err := viper.Unmarshal(&config, viper.DecodeHook(
+		mapstructure.ComposeDecodeHookFunc(
+			mapstructure.StringToSliceHookFunc(","),
+			viperDecodeHookFunc(),
+		))); err != nil {
+		return nil, err
+	}
+
+	return &config, nil
+}
+
+// viperDecodeHookFunc creates a custom decode hook to handle the Plugins field.
+func viperDecodeHookFunc() mapstructure.DecodeHookFunc {
+	return func(f reflect.Type, t reflect.Type, data interface{}) (interface{}, error) {
+		// Default handling for non-Plugin fields
+		if t != reflect.TypeOf([]Plugin(nil)) {
+			return data, nil
+		}
+
+		// Cast the plugins data to a slice of interfaces
+		pluginsDataSlice, ok := data.([]interface{})
+		if !ok {
+			return nil, errors.New("invalid plugins field")
+		}
+
+		plugins := make([]Plugin, 0)
+
+		// Iterate over the plugins data and decode each plugin
+		for _, pluginData := range pluginsDataSlice {
+			// Cast the plugins map to a map[string]interface{}
+			pluginsMap, ok := pluginData.(map[string]interface{})
+			if !ok {
+				return nil, errors.New("invalid plugins field")
+			}
+
+			name, ok := pluginsMap["name"].(string)
+			if !ok {
+				return nil, fmt.Errorf("plugin name missing or not a string")
+			}
+
+			plugin := PluginFactory(name)
+			if plugin == nil {
+				return nil, fmt.Errorf("unsupported plugin type: %s", name)
+			}
+
+			if err := mapstructure.Decode(pluginsMap, plugin); err != nil {
+				return nil, err
+			}
+
+			plugins = append(plugins, plugin)
+		}
+
+		return plugins, nil
+	}
+}
+
+// OrderBookPluginConfig encapsulates the order book plugin configuration.
+type OrderBookPluginConfig struct {
+	Enabled bool   `mapstructure:"enabled"`
+	Name    string `mapstructure:"name"`
+}
+
+// GetName implements Plugin.
+func (o *OrderBookPluginConfig) GetName() string {
+	return o.Name
+}
+
+// IsEnabled implements Plugins.
+func (o *OrderBookPluginConfig) IsEnabled() bool {
+	return o.Enabled
+}
+
+var _ Plugin = &OrderBookPluginConfig{}
 
 type EndpointOTELConfig struct {
 	Quote float64 `mapstructure:"/router/quote"`
@@ -115,4 +205,15 @@ func validateDynamicMinLiquidityCapDesc(values []DynamicMinLiquidityCapFilterEnt
 	}
 
 	return nil
+}
+
+// PluginFactory creates a Plugin instance based on the provided name.
+func PluginFactory(name string) Plugin {
+	switch name {
+	case orderbookplugindomain.OrderBookPluginName:
+		return &OrderBookPluginConfig{}
+	// Add cases for other plugins as needed
+	default:
+		return nil
+	}
 }