+++ router/usecase/router_usecase_test.go
@@ -722,8 +722,7 @@ func (s *RouterTestSuite) TestGetCandidateRoutes_Chain_FindUnsupportedRoutes() {
 	s.Require().NoError(err)
 
 	// Unmarshal the config into your Config struct
-	var config domain.Config
-	err = viper.Unmarshal(&config)
+	config, err := domain.UnmarshalConfig()
 	s.Require().NoError(err)
 
 	// Set up mainnet mock state.
@@ -829,8 +828,7 @@ func (s *RouterTestSuite) TestPriceImpactRoute_Fractions() {
 	s.Require().NoError(err)
 
 	// Unmarshal the config into your Config struct
-	var config domain.Config
-	err = viper.Unmarshal(&config)
+	config, err := domain.UnmarshalConfig()
 	s.Require().NoError(err)
 
 	// Set up mainnet mock state.
@@ -1359,7 +1357,7 @@ func (s *RouterTestSuite) TestGetCustomQuote_GetCustomDirectQuotes_Mainnet_Order
 			poolID:        1904,
 			err: domain.OrderbookNotEnoughLiquidityToCompleteSwapError{
 				PoolId:   1904,
-				AmountIn: sdk.NewCoin(NATIVE_WBTC, osmomath.NewInt(999000000000)),
+				AmountIn: sdk.NewCoin(NATIVE_WBTC, osmomath.NewInt(999000000000)).String(),
 			},
 		},
 	}